moving collapsibility from subform to property-group
